pyproject.toml
[project]
name = "goodreads-ml-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.2",
    "pydantic>=2.12.5",
    "pyyaml==6.0.2",
    "scrapy>=2.14.1",
]
